adk-gemini-cli-workshop/pyproject.toml
[project]
name = "simple-travel-agent-workshop"
version = "0.2.0"
description = "Simple Travel Agent Workshop using Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk",
    "mcp",
]
